fix(artifactresolver) Overhaul exception handling in artifact resolver
Problem:
When building the e2e tests, a lot of stack traces occur. This is due to
some ArtifactNotFound exceptions which are thrown during normal
execution of the artifact resolver. These are expected exceptions so
they should not be shown to the user.

Mitigation:
Overhaul exception strategy: Instead of throwing and recatching
exceptions, we now directly catch exceptions and log only the relevant
information. "requestFile" in IArtifactRequester may now return an empty
File. This will (as before) be communicated to the user anyway as
artifacts would be missing from the final report. Exceptions are only
thrown if the build is not recoverable.

Signed-off-by: Martin Idel <external.Martin.Idel@bosch-si.com>