fix(p2resolver): Better error reporting
The resolver uses a headless eclipse product.
The product is shipped with three launchers for three platforms:
- MacOS with cocoa (x86_64)
- Linux with GTK (x86_64)
- Windows (x86_64)
The launcher is a shared library, which on Linux is explicitly linked
against glibc. Some Linux/Unix distributions however are shipped
without glibc. This is only visible during runtime and this commit
improves the error presented to the user.

Signed-off-by: Martin Idel <external.Martin.Idel@bosch-si.com>