change condition for checking for git diff in writeIndexRootFile
Since the Jenkins job failed when trying to checkout to current branch
because changes in branch might be lost, condition must not have worked.

Hence this changes the if condition to decide if a commit should happen.
`--quiet` was added to stop the search as soon as a difference has been
found. The command implies the git parameter `exit-code` and will exit
with 1 when the tree is modified
`--` was added for the command to work as intended, since it separates
the path from the rest of the argument and helps Git know that `HEAD`
is not the name of a file.
The condition is specifically designed to exit with an error if the
script is using `set -e`, as described here:
https://stackoverflow.com/a/28296286

Signed-off-by: Stephanie Neubauer <Stephanie.Neubauer@bosch-si.com>