fix path function sw360 release
This changes the url that was given to the RestTemplate that
executes the Patch function.
Opposed to the create function that is done with a POST request, the
PATCH function needs the releaseId in the url, otherwise a 405
Method Not Allowed Error is thrown.

The if-statement that checks if the Update_releases flag is set
and then throws and Exception was removed, since it is no longer
necessary.
The documentation is changed as well, since the mention of
feature not being supported is no longer accurate.

Signed-off-by: Stephanie Neubauer <Stephanie.Neubauer@bosch-si.com>