bug fix #349 missing component type in create component request
When creating a component for a new release that has no component in
SW360 yet the post request failed with a 400 null BadRequest error.
This was due to missing required value "componentType".
This was not set since the component is created from the release.
The release does not contain that information.

Hence, a variable was introduced to the release that has a
JsonIgnore property to set an isProprietary boolean value.
This allows to check for the value when creating a component from
release.

It is not ideal, since otherwise the new variable is not used or
necessary and creates unnecessary overhead.

Signed-off-by: Stephanie Neubauer <Stephanie.Neubauer@bosch-si.com>