make uploading attachments that are source files in artifact possible
Activates and refactors feature to upload attachments that are of type
SOURE to SW360 for releases.
This has been separated from the getOrCreateRelease() function for
the purpose of separating the two functionalities of creating or updating
a release with release information and uploading files as attachments
that are linked to the release. This also reduces the necessity to carry
object information of the attachments through several layers and functions.
For this a SW360AttachmentAdapterUtils class was created to extract
attachment information from the Artifact.

Signed-off-by: Stephanie Neubauer <Stephanie.Neubauer@bosch-si.com>