Add download attachment feature to enricher
This includes a feature for the enricher to be able to download
attachments, specifically source files from release, which currently
is the only attachment information we want. This can be extended.
For this two configuration variables have been added to the enricher.
A boolean that marks if the download should occur and a string that
gives the path the attachments should be downloaded to.
In order to be able to deserialize the needed attachment information
new classes for SW360Attachment List, ListEmbedded and Sparse are
written.
Currently the downloading client functionalities are done via the
ReleaseClient, since it inherits the abstract AttachemntAwareClient.
This could be potentially generalized, should it every be necessary,
however, currently only the releases need the download functionality.
Only attachments that are of type SOURCE currently get downloaded since
they are the only ones we want to map.

Signed-off-by: Stephanie Neubauer <Stephanie.Neubauer@bosch-si.com>