Rework exception handling in Sw360 module
Fixes #358
Exception handling was reworked to catch places where due to a
runtime excpetion thrown within SW360 the Antenna build failed
without throwing an ExecutionException and without a log/excpetion
explaining the error and source properly.

Furthermore most thrown ExecutionExceptions where removed and
replaced with LOGGER inputs to avoid the Antenna build to fail.
Now an error or debug message will be logged.
Errors in the communication with SW360 should not make the Antenna
run fail, since they are not crucial for an attribution document.

The process of checking the ResponseEntity has also been simplified.
In a check respone function the status code is checked and an
ExecutionExcpetion is thrown that is caught later to have a
Logger entry.
The checks if the Body is null has been moved to Utils as well.
An item is given to a function and checked for null and then
a Getter get method is invoked on a follow up function. If
the body is null an IllegalStateException is thrown.

This structure separates the exception handling from the code more.

The function getProject in the SW360ProjectClient and the test
testing that function are removed because they are never used.

Signed-off-by: Stephanie Neubauer <Stephanie.Neubauer@bosch-si.com>