integrate ORT Downloader as a processor step Resolves https://github.com/eclipse/antenna/issues/326
To download the source files of artifacts, which don't have one yet,
the ORT Downloader is used. Therefore a new processor step is
implemented which uses the Downloader class to download the sources
to a tmp directory, archieve it in a ZIP file and store it in the
antenna model.

Signed-off-by: Onur Demirci <Onur.Demirci@Bosch-si.com>