Restructure serialization of SW360HalResources
This uses the object mapper from the jackson databind library to
serialize and deserialize all SW360Objects except the attachments
(they are not compatible with the format yet and this should be done
when the roundtrip pr is merged).
This is adapted to the way the releases already were serialized and
helps make the code cleaner by avoiding set attribute variables in
a separate file and instead have the serialization with the objects.
Should the components or projects get externalIds those will also
be easier to map and test.

Tests have been added for the serialization of the objects.

SW360 Licenses are not serialized with the generic function yet because
of uncertainty with the full name short name situation.

Signed-off-by: Stephanie Neubauer <Stephanie.Neubauer@bosch-si.com>