Adding attachment to release serialization
This enables to get the list of attachments a release has when
getting the release itself from SW360. Thus making it possible
to check for duplicate names in the attachments without additional
requests.

Some new tests had to be written for the SW360ReleaseTest and
were consequently implemented in all test classes extending the
SW360ResourcesTestUtils.

Signed-off-by: Stephanie Neubauer <Stephanie.Neubauer@bosch-si.com>