Add new cli frontend for compliance tool with sw360 exporter.
This introduces a new cli frontend that uses the antenna sw360
classes to create a csv file with all non cleared/ approved releases
in a given sw360 instance.

This is the start of a frontend for a compliance tool that can be
used to manage metadata in a given sw360 instance, enabling to get
a list of all releases that need approval to be able to use that
information for curation processes.

The test is set to ignored since it requires an actual sw360 instance.

Signed-off-by: Stephanie Neubauer <Stephanie.Neubauer@bosch-si.com>