Refactoring and adding tests
Extracts init of SW360 into main class so it is decoupled from
the Configuration and Adapter classes. The execute returns an
int to give the exit code to the main. Thus no System.exit
function is called outside main.
The initialization of the SW360Exporter in main is now tested.

Adds SW360configurationTest and some other minor refactoring

Signed-off-by: Stephanie Neubauer <Stephanie.Neubauer@bosch-si.com>