basic version of the sw360 updater
This introduces the sw360 updater functionality to the
compliance tool. This takes the same csv file format
that is also produced by the exporter and updates the
release information added in the csv file to a given
sw360 instance.

This is the foundatoin for making this feature capable of
approving releases.

This PR includes tests and documentation

Signed-off-by: Stephanie Neubauer <Stephanie.Neubauer@bosch.io>