Refactor CsvAnalyzerImpl to a more specific csvreader class
Issue: #404

Move CsvAnalyzerImpl functionality into a new class CSVReader.
This prepares to have a roundtrip csv read and write
functionality in one class, which minimizes the danger of
adding or removing a column to one of the functions without
also adapting the other one.

Tests are done through testing the CSVAnalyzer.

Updates other usages of the CsvAnalyzerImpl to use CSVReader

Signed-off-by: Stephanie Neubauer <Stephanie.Neubauer@bosch.io>