SW360 Map Release To Artifact
Map release to artifact in the ReleaseAdapterUtils.
Consequently adapting the EnricherImpl to use the
utils method.

Question of how to map downloads to artifacts in utils class
remains. Technically they belong there as a fact, but need a
clients to actually download the fact.

Hence the roundtrip with files remains incomplete.

Signed-off-by: Stephanie Neubauer <Stephanie.Neubauer@bosch.io>