Ensure SW360 has a roundtrip artifact -> release
Issue: #405

For this a test class is added that asserts for roundtrips.
Additionally the TODO is removed from the AdapterUtils. The
sources can't be part of this roundtrip since they require
a call to an actual existing SW360 instance and no indempotent
function can thus be created.

If one also wants sources in the round trip one has to add
them after converting the first object into the other.

Functions for this exist in the Enricher and the Updater.
It should not be put anywhere deeper, since we want to keep
all the client classes free from Artifact dependencies.

Error with wrong package path in updater tests fixed.

Signed-off-by: Stephanie Neubauer <Stephanie.Neubauer@bosch.io>