Changed assembly plugin configuration in frontend-stubs-testing.
Issue: eclipse#383

Previously, the project produced a single artifact which was a fat jar
including all dependencies. This could cause problems in projects that
dependent on this jar, especially with multiple sl4fj bindings. The
changed configuration now causes two artifacts to be generated: the
plain jar (which is now used by client projects) and the fat jar,
which is needed for integration testing.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>