Enabled a mockito extension for mocking of final methods.
Issue #397.

Mockito per default does not support mocking final classes or methods.
By using this extension configuration, final mocking can be made to
work.

While mocking of final methods should be used carefully in general,
there are libraries that make testing without this mechanism almost
impossible. In the concrete case, the correct configuration of an
Apache HTTP client instance should be tested; but the class does not
expose methods to query its properties nor can the
corresponding builder class be mocked as the relevant methods are
final.

The extension of the marker file has also been excluded in the script
that checks the license headers.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>