Improved handling of rest template in SW360Client.
Issue #397.

A constructor was added that accepts the template. The template is
expected to be correctly configured and is used for all requests
(including PATCH which had been handled in a special way). Access to
the template field has been encapsulated. A test class was added that
checks basic REST calls.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>