Remove Drools Policy Engine
Removal of the drools module with three submodules and the common-rules package.
The drools engine has been removed by the Java based policy engine. After an evaluation of the
result of the drools engine, we decided to not push the drools idea further since it
did not bring advantages for the rule developers. The idea with the bdd driven
specification of rules is still followed but with the new Java based policy engine.
Common-rules was a simple package of rules that we could publish as open source, a
corresponding package is also available in the Java policy engine.

Signed-off-by: Lars Geyer-Blaumeiser <lars.geyer-blaumeiser@bosch.io>