Remove P2 functionality not needed anymore
Remove ChildJarResolver. We looked into the functionality, it was copying some weird source jars, but the functionality was not
understood anymore. So it seems not to be important and we remove it until further notice.
We removed the p2 resolver and the p2 product because it is a very special way of downloading sources which will not be needed
anymore with the way of working in SW360. Therefore we removed the complexity of that from the tooling.
Since this resulted in only one element left in the p2 package which could be also seen as something relevant for Maven,
the ManifestResolver has been moved to the Maven module.
As a result some tests needed to be changed to adapt to the other changes and the build had to be simplified because
the Eclipse product build is removed from the build workflow. All potential stuff that is not needed anymore, has been
removed.

Signed-off-by: Lars Geyer-Blaumeiser <lars.geyer-blaumeiser@bosch.io>