Fixed problem with build directories in ort pom.
Issue: eclipse#428

The build section of the pom contained explicit declarations for the
sources and test sources directories. These are unnecessary because
all the paths to be compiled are referenced again in the configuration
of the Kotlin plugin. In addition, these declarations had the side
effect that the standard paths of the maven compiler plugin were
changed, too. So the test java class was not picked up during test
compilation.

After removing these declarations, all unit tests are now executed
during a Maven build.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>