Removed SW360 user client and adapter classes.
Issue: eclipse#423

The classes were not used in the codebase, and their sole
functionality, querying a user by an email address, was dysfunctional.
(The REST call used for this purpose is not working.)

To have a comprehensive SW360 client library, it may be necessary to
re-introduce the classes later again; but then a different API and
design will be used.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>