Add test to SW360*ClientAdapter
Add test to SW360ProjectClientAdapter:

This brings a refactoring that separates the setting of the
project client withing the project client adapter to a separate
set method. Since it should only be set once a null check was
introduced in the setter.

The method "hasEqualCoordinates" was extracted to the utils,
since this is the proper place for it. Tests were added to
the adapter utils tests.

Add test to SW360LicenseClientAdapter:

This brings a refactoring that separates the setting of the
license client withing the license client adapter to a separate
set method.

Add test to SW360ComponentClientAdapter:

This brings a refactoring that separates the setting of the
component client withing the component client adapter to a
separate set method.

Add test to SW360ReleaseClientAdapter

This brings a refactoring that separates the setting of the
release client withing the release client adapter to a
separate set method. The component client is still set
via the constructor

Signed-off-by: Stephanie Neubauer <Stephanie.Neubauer@bosch.io>