Add test to SW360ProjectClientAdapter
This brings a refactoring that separates the setting of the
project client withing the project client adapter to a separate
set method. Since it should only be set once a null check was
introduced in the setter.

The method "hasEqualCoordinates" was extracted to the utils,
since this is the proper place for it. Tests were added to
the adapter utils tests.

Signed-off-by: Stephanie Neubauer <Stephanie.Neubauer@bosch.io>