Some improvements of ProxySettings class.
Issue: eclipse#423

The class no longer stores a flag whether a proxy should be used or
not. Rather, this information is derived from the other settings
(which have invalid values if no proxy should be used). This is
possible because in case that no proxy is used, the settings for proxy
host and port are irrelevant.

Added factory methods for creating instances based on different use
cases. Client classes have been adapted accordingly.

Added a test dependency to the equals-verifier library to have a
better way of testing equals() and hashCode() implementations.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>