Adds default value of product name and version to doc generator
Issue: #437

This enables the documentation generator to get the source of
their product name and version from the tool config if it has
not been set.

This also means that both the document generator and the
sw360 updater can access the same source to find the name
of the product and its version.

Signed-off-by: Stephanie Neubauer <Stephanie.Neubauer@bosch.io>