Reworked SW360LicenseClient to use the HTTP support library.
Issue: eclipse#423

The client class now sends asynchronous requests and returns futures
with the results. The error handling has been slightly modified; a
request for a non-existing license no longer returns an empty
optional, but a failed future. (Antenna-specific error and fallback
handling now needs to be implemented on a different layer.)

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>