pom: Add a comment about the compile scope of testing modules
Strictly speaking, the explicit compile scope would only be needed for
dependencies that actually are declared in any of the parent POMs, i.e.
those that lack a version in the POM changed by this commit, but making
that distinction would be too confusing, to just be explicit about all
compile dependencies.

Signed-off-by: Sebastian Schuberth <sebastian.schuberth@bosch.io>