Added central service objects to AntennaContext.
Issue: eclipse#423

The HTTP client and the JSON object mapper are thread-safe objects
whose creation is expensive. Therefore, it makes sense to have them
created once at a central place and to provide access to them to all
components that require them. Currently, such objects are created at
multiple places where they are used; this will be changed to make use
of the central service objects.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>