Adapted the compliance tool to changes on the SW360 configuration.
Issue: eclipse#423

In order to create an SW360 configuration, the HTTP client and the
JSON mapper now needs to be passed in; so these objects have to be
created outside. To avoid redundancy, the compliance tool reuses the
AntennaContext class for the creation of these objects; it creates a
partially initialized context from its own configuration.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>