Added ServiceFactory class.
Issue: eclipse#423

This class implements the creation of the shared service objects like
the JSON mapper and the HTTP client. It is going to be used by the
Antenna context. The goal is to use this class for the service
creation in the compliance tool rather than a full-blown Antenna
context.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>