HttpHelperTest: Download to a different location that the source file
Previously, the downloaded file was written to the same location from
where the file was served. This coincidentally used to work before
63781ba because FileUtils.copyInputStreamToFile() truncated an existing
file instead of deleting it like Files.copy() does, but with 63781ba the
test started to fail on Windows.

Also, the created test file was empty, and the final assertion only
compared the file paths, not their contents.

Improve this by

- using separate directories for source and target files,
- writing some content to the test file,
- compare the contents of source and target files.

Signed-off-by: Sebastian Schuberth <sebastian.schuberth@bosch.io>