Upgrade ORT to the latest version
The upgrade is required to be able read the changed ORT result files.
Several minor API changes need to be accounted for, as well as change of
package names due to the move of ORT to its own GitHub organization [1].

Also, this version of ORT includes a fix that resolves a Gradle vs. Maven
version conflict resolution issue regarding the "jsch" dependency [2].

[1] https://github.com/oss-review-toolkit/ort
[2] https://github.com/oss-review-toolkit/ort/pull/2440

Signed-off-by: Sebastian Schuberth <sebastian.schuberth@bosch.io>