Add "/token" to all authentication urls
Since the suffix "/token" is not added automatically
to the authentication url in the sw360authentication
client anymore, it needs to be added to the authentication
url in the configuration to work with an out of the box
locally set up sw360.

Changes were made in minimal example project, example
project, ExampleTestProject and the documentation.

Signed-off-by: Stephanie <Stephanie.Neubauer@bosch.io>