Improved documentation how to obtain an HttpClient.
Issue: eclipse#452.

The documentation of the HTTP support library no longer refers to the
AntennaContext class; rather, the creation of the client via the
corresponding factory interfaces is described in detail.

The Antenna development documentation has been extended to mention
that shared service objects can be obtained from the AntennaContext.

Signed-off-by: Oliver Heger <60134199+oheger-bosch@users.noreply.github.com>