Fixed problems with equals() methods in SW360 resources.
Issue: eclipse#426.

The current implementations, due to the inheritance from
SW360HalResource, violated the symmetry conditions of the equals()
contract. This has been fixed by introducing a canEqual() method.
Also, the resource classes have been made final; this was for two
reasons: EqualsVerifier was complaining because equals() could be
overridden by a subclass, and for plain data classes it makes sense to
be final. A dedicated test class for equals() methods has been
introduced, as for many classes the equals() logic was not tested.

SW360HalResource now directly supports querying the ID of a resource.
Previously, some resource classes had redundant implementations for
that purpose.

Signed-off-by: Oliver Heger <60134199+oheger-bosch@users.noreply.github.com>