Close all opened PDDocument objects
Fixes https://github.com/eclipse/antenna/issues/490

The class Templates wrapped all PDDocuments, which are needed for the
generation of the Attribution Document. After opening the PDDocuments
during the initialization of the Templates object, they were not
closed again which caused the Warning described in the issue. After
closing all PDDocument objects explicitly, the warning disappeared.

Signed-off-by: Onur Demirci <onur.demirci@bosch.io>