Renamed documentation file for HTTP support library.
Issue: eclipse#506.

When building the jar with dependencies for the Antenna CLI this file,
due to the name clash, was overriding the documentation for the CLI;
so instead of the CLI documentation the documentation for the HTTP
support library was displayed. Renaming the file avoids the name
clash; in general, to avoid such problems, documentation files for
modules referenced by other artifacts should have unique names.

Signed-off-by: Oliver Heger <60134199+oheger-bosch@users.noreply.github.com>