Fixed CSVBasedLicenseKnowledgeBase with unknown license names
The CSVBasedLicenseKnowledgeBase does not behave the same way as the
SPDXLicenseKnowledgeBase i.e. in cases that the license name is
unknown, it was kept empty instead of setting the with the license id.
This fixes also the anchor error in the Attribution Document Generator,
which was not able to create an anchor of an empty string.

Signed-off-by: Onur Demirci <onur.demirci@bosch.io>