SW360ClientConfig now stores the base REST URI.
Issue: eclipse#426.

This URI is going to be used to resolve relative URIs against it. This
is used as a work-around for a problem with the self links returned by
SW360: these are absolute URLs, which are not necessarily correct when
the server is behind a proxy or load balancer.

Signed-off-by: Oliver Heger <60134199+oheger-bosch@users.noreply.github.com>