Request release from sw360 only when clearing status not INITAL
Issue: #529

This allows viewer calls to the SW360 instance that are
unnecessary since nothing will be done with the fetched
information later on.

To enable this the function in the UpdaterImpl was split
into two, one in which only the artifact is mapped onto
the release with the information of the artifact and then
calls the second method mapping release information from
an SW360 instance on the release.
That method can be called independently of the first.
Allowing to give a release object that was not created
freshly in the method.

The tests were adapted.

Signed-off-by: Stephanie Neubauer <Stephanie.Neubauer@bosch.io>