Made attachment download more generic and flexible.
Issue: eclipse#538.

SW360AttachmentAwareClient now offers a method to process an
attachment that expects a generic processor object. The processor is
passed a stream with the content of the attachment. This method is
going to replace the download method (after all references have been
removed).

While the download method served a specific use case, the new approach
is much more flexible. For instance, a processor can decide where and
under which file name the attachment is saved; or whether it is saved
at all (maybe only a checksum has to be calculated or the data needs
to be transferred to somewhere else).

Signed-off-by: Oliver Heger <60134199+oheger-bosch@users.noreply.github.com>