Made access to source attachment more consistent.
Issue: eclipse#538.

The Artifact model of Antenna supports only a single source
attachment. Therefore, it does not really make sense to add this
single attachment to a collection and treat it as multiple
attachments; this only increases complexity.

Changed ArtifactToAttachmentUtils.getSourceAttachmentFromArtifact() to
return an Optional with the single attachment path.

Signed-off-by: Oliver Heger <60134199+oheger-bosch@users.noreply.github.com>