Add the open source font NotoSans in attribution-document-basic-bundle
The AttributionDocumentGenerator uses the Times-Roman font if no font is
provided in the template bundle. To avoid the use of a default font
which doesn't support some glyphs, the open source font NotoSans is
added to the basic template bundle. In addition, the template PDF files
were edited with the new font and re-exported.

Signed-off-by: Onur Demirci <onur.demirci@bosch.io>