Enhancements of SW360MetaDataUpdater regarding uploads.
Issue: eclipse#538.

The method to upload attachments now has an improved handling of
conflicts. By specifying a flag it can be controlled whether conflicts
(already existing attachments) are reported as errors or tried to be
resolved by deleting the conflicting attachment. The method now
returns an upload result, so that callers can deal with partially
successful uploads.

There is now an extended method to delete attachments which accepts a
predicate; so the caller has full control over the attachments to be
deleted.

Signed-off-by: Oliver Heger <60134199+oheger-bosch@users.noreply.github.com>