Enhancements of SW360UpdaterImpl regarding uploads.
Issue: eclipse#538.

When creating a release entity from an artifact, it is now possible to
upload additional files. This will be used for the clearing report.

When deleting obsolete source attachments (if enabled), the current
source attachment is not deleted. It need not be touched unless a
conflict is detected.

Signed-off-by: Oliver Heger <60134199+oheger-bosch@users.noreply.github.com>