For failed responses the entity is now read.
Issue: eclipse#558.

FailedRequestException now has a property for the error message sent
by the server. When creating an exception from a response indicating a
failed request, the body stream of the response is now read, and the
resulting message is stored in the exception. This information can be
helpful for callers to figure out what actually went wrong.

Signed-off-by: Oliver Heger <60134199+oheger-bosch@users.noreply.github.com>