Working "docker run"
Issue: #365

Base image was upgraded to a higher version.

Added batch mode to Maven as explained here
https://books.sonatype.com/mvnref-book/reference/running-sect-options.html
These is essential to run Maven in a non-interactive, continuous
integration envionment (such as Docker).

The removal of the antenna and maven folder is not necessary
since the base image will not be included in the final
image and therefore reducing size this way is not possible.

changes in cmd, "-" before "jar" was missing

Signed-off-by: Stephanie Neubauer <Stephanie.Neubauer@bosch.io>