Fix a problem with a platform-specific path conversion.
Issue: eclipse#591.

CSVArtifactMapper tries to convert paths to source files to relative
paths, so that they are not specific to a local configuration.
However, for absolute paths that are not below the base directory,
this obviously does not work reliably on all platforms - and does not
make any sense either. So such paths are now stored as absolute paths.

Resolves #591

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>