Clearing document name contains name and version of release now
Resolves #587.

Before, the name of the clearing document, if it was generated,
consisted of the release's ID. Due to a change in the logic of the
`SW360Update`, a release did not have an ID when the document was
generated, resulting in a name containing `null`. To circumvent this and
to make it more readable, the name is now combined by the name and
version of a release.

Signed-off-by: Korbinian Singhammer <external.Korbinian.Singhammer2@bosch.io>