Update docker documentation for usage of properties file
Resolves #584

The Antenna CLI is already capable of using a properties file.
Therefore, update the documentation how to use such a file within the
docker run command. This is done by either having it inside of the
project folder that is mounted on the docker daemon or by including it
with the -v (--volume) option.

Signed-off-by: Korbinian Singhammer <external.Korbinian.Singhammer2@bosch.io>