Replace java.util.zip with apache.commons.compress.archiver.zip
Resolves #601

java.utils.zip throws an exception when zipping a jar with duplicate
entries. This is not the case for the the compress package. Thus, the
java standard is replaced. Furthermore, ZipExceptions are not caught
anymore, since they are not pressent in the compress package and are
also derived from IOExcetption but are already taken care of.

Signed-off-by: Korbinian Singhammer <external.Korbinian.Singhammer2@bosch.io>