Replace '/' and '\' by '_' in auto generated clearing document name
Resolves #615

When no clearing document is specified for a release to be updated, a
document is generated with a name consisting of the release's name and
version. However, this name might contain a '/' (linux) or '\' (windows)
character and the documents name is interpreted to contain a subdirectory,
leading to file errors. Thus, every '/' and '\'  occuring is replaced by
a '_' character.

Signed-off-by: Korbinian Singhammer <external.Korbinian.Singhammer2@bosch.io>