Fix wip not overwritting release information
It was discovered that a release would not update already existing
information of a release already in SW360 due to the mergeWith
method taking the SW360 release as the object of precedence.

For the use cases so far this is desirable, as they see SW360 as
the source of truth. But the WIP state in the compliance tool
is designed specifically to adjust and change existing information.

Hence this commit introduces a way of using the WIP state in the
compliance tool by introducing an "artifactHasPrecedence" flag
to the getOrCreateRelease method that decides which release
has precedence in the merge.

Resolves #588

Signed-off-by: Stephanie Neubauer <Stephanie.Neubauer@bosch.io>