Add order to clearing state to avoid overwritting approved data
Since the WIP state could be used to overwrite already cleared
information, an additional condition was introduced to merge
the release gotten from the artifact as precedence.

The clearing state of the artifactneeds to be higher or equal
to the release it wants to merge precendence with

Documentation was added for the new functionality.

Resolves #588

Signed-off-by: Stephanie Neubauer <Stephanie.Neubauer@bosch.io>